used uint64 in headers for encoding type sizes

+++ addrcompressor/lut.go
@@ -37,8 +37,8 @@ type CompressionLut struct {
 	IdxToKeyLut []string
 	dict        *ccmap.ConcurrentMap
 	tempLut     *CompressionLut
-	length      uint32
-	offset      uint32
+	length      uint64
+	offset      uint64
 	lock        sync.RWMutex
 	depths      [][2]int
 }
